+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/Teleop.java
@@ -3,9 +3,11 @@ package org.firstinspires.ftc.teamcode.opmodes;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.Range;
 
+import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.Odometry;
 import org.firstinspires.ftc.teamcode.Robot;
 import org.firstinspires.ftc.teamcode.toolkit.MathFunctions;
+import org.firstinspires.ftc.teamcode.toolkit.TelemetryOutput;
 import org.firstinspires.ftc.teamcode.toolkit.ULLinearOpMode;
 
 @TeleOp(name = "TeleOp", group = "OpModes")
@@ -26,6 +28,7 @@ public class Teleop extends ULLinearOpMode {
         waitForStart();
         robot = new Robot();
         odom = new Odometry(robot);
+        odom.run();
 
         while(opModeIsActive()) {
 
@@ -49,15 +52,7 @@ public class Teleop extends ULLinearOpMode {
             robot.drive(magnitude, joystickAngle, turnValue);
 
             // add telemetry data for the encoders
-            telemetry.addData("Left Encoder pos:\t", odom.getLeftTicks() / Robot.COUNTS_PER_INCH);
-            telemetry.addData("Right Encoder pos:\t", odom.getRightTicks() / Robot.COUNTS_PER_INCH);
-            telemetry.addData("Center Encoder pos:\t", odom.getCenterTicks() / Robot.COUNTS_PER_INCH);
-            telemetry.addData("LeftDelta:\t", odom.deltaLeftDistance);
-            telemetry.addData("RightDelta:\t", odom.deltaRightDistance );
-            telemetry.addData("WorldX:\t", MathFunctions.truncate(odom.worldXPosition));
-            telemetry.addData("WorldY:\t", MathFunctions.truncate(odom.worldYPosition));
-            telemetry.addData("WorldOrientationAngle\t", Math.toDegrees(odom.worldAngle));
-            telemetry.update();
+            TelemetryOutput.printFullTelemetry(telemetry, odom);
 
         }
